Separate user for console sessions
The user that will be used for console sessions can be configured in the
agent's config.yaml and defaults to a 'console' user by default. This
default user is created upon RPM installation (if it doesn't exist) and
is added to the 'wheel' supplimental group so that it gets passwordless
sudo access.

The SELinux domain for the console shell is transitioned to the normal
`unconfined_t` domain which is used by normal shell users so that things
are more consistent with a normal shell session (e.g. via ssh).